+++ tests/Unit/Public/Add-PstProject.tests.ps1
@@ -1,15 +1,20 @@
 
-Describe "$commandName Unit Tests" -Tag 'UnitTests' {
+Describe "Add-PstProject Unit Tests" -Tag 'UnitTests' {
     Context "Validate parameters" {
         It "Should only contain our specific parameters" {
-            $command = 'Restore-PstDay'
+            $command = 'Add-PstProject'
             [object[]]$params = (Get-Command $command).Parameters.Keys
-            [object[]]$specificParameters = 'Date'
+            [object[]]$specificParameters = 'Client','Project'
             $allCommonParameters = [System.Management.Automation.PSCmdlet]::CommonParameters + [System.Management.Automation.PSCmdlet]::OptionalCommonParameters
             
             # Filter out common parameters to check only specific ones
             $actualSpecificParams = $params | Where-Object { $_ -notin $allCommonParameters }
             (@(Compare-Object -ReferenceObject $specificParameters -DifferenceObject $actualSpecificParams).Count ) | Should -Be 0
+        }
+
+        It "Should have mandatory parameters" {
+            (Get-Command Add-PstProject).Parameters['Client'].Attributes.Mandatory | Should -Be $true
+            (Get-Command Add-PstProject).Parameters['Project'].Attributes.Mandatory | Should -Be $true
         }
     }
 }