+++ tests/Unit/Public/Add-PstTime.tests.ps1
@@ -3,11 +3,13 @@ Describe "$commandName Unit Tests" -Tag 'UnitTests' {
     Context "Validate parameters" {
         It "Should only contain our specific parameters" {
             $command = 'Add-PstTime'
-            [object[]]$params = (Get-Command $command).Parameters.Keys | Where-Object {$_ -notin ('whatif', 'confirm')}
-            [object[]]$knownParameters = 'Client','Project','Minutes','StartTime'
-            #$knownParameters += [System.Management.Automation.PSCmdlet]::CommonParameters
-
-            (@(Compare-Object -ReferenceObject ($knownParameters | Where-Object {$_}) -DifferenceObject $params).Count ) | Should -Be 0
+            [object[]]$params = (Get-Command $command).Parameters.Keys
+            [object[]]$specificParameters = 'Client','Project','Minutes','StartTime'
+            $allCommonParameters = [System.Management.Automation.PSCmdlet]::CommonParameters + [System.Management.Automation.PSCmdlet]::OptionalCommonParameters
+            
+            # Filter out common parameters to check only specific ones
+            $actualSpecificParams = $params | Where-Object { $_ -notin $allCommonParameters }
+            (@(Compare-Object -ReferenceObject $specificParameters -DifferenceObject $actualSpecificParams).Count ) | Should -Be 0
         }
     }
 }